+++ webtest/forms.py
@@ -112,10 +112,19 @@ class Select(Field):
         """
         self._forced_value = value
 
+    def get_value_for_text(self, text):
+        for i, (option_value, checked, option_text) in enumerate(self.options):
+            if option_text == utils.stringify(text):
+                return option_value
+
+        raise ValueError("Option with text %r not found (from %s)"
+                         % (text, ', '.join(
+                             [repr(t) for o, c, t in self.options])))
+
     def value__set(self, value):
         if self._forced_value is not NoValue:
             self._forced_value = NoValue
-        for i, (option, checked) in enumerate(self.options):
+        for i, (option, checked, text) in enumerate(self.options):
             if option == utils.stringify(value):
                 self.selectedIndex = i
                 break
@@ -123,7 +132,7 @@ class Select(Field):
             raise ValueError(
                 "Option %r not found (from %s)"
                 % (value, ', '.join(
-                [repr(o) for o, c in self.options])))
+                [repr(o) for o, c, t in self.options])))
 
     def value__get(self):
         if self._forced_value is not NoValue:
@@ -131,7 +140,7 @@ class Select(Field):
         elif self.selectedIndex is not None:
             return self.options[self.selectedIndex][0]
         else:
-            for option, checked in self.options:
+            for option, checked, text in self.options:
                 if checked:
                     return option
             else:
@@ -158,10 +167,26 @@ class MultipleSelect(Field):
         self._forced_values = values
         self.selectedIndices = []
 
+    def get_value_for_texts(self, texts):
+        str_texts = [utils.stringify(text) for text in texts]
+        value = []
+        for i, (option, checked, text) in enumerate(self.options):
+            if text in str_texts:
+                value.append(option)
+                str_texts.remove(text)
+
+        if str_texts:
+            raise ValueError(
+                "Option(s) %r not found (from %s)"
+                % (', '.join(str_texts),
+                   ', '.join([repr(t) for o, c, t in self.options])))
+
+        return value
+
     def value__set(self, values):
         str_values = [utils.stringify(value) for value in values]
         self.selectedIndices = []
-        for i, (option, checked) in enumerate(self.options):
+        for i, (option, checked, text) in enumerate(self.options):
             if option in str_values:
                 self.selectedIndices.append(i)
                 str_values.remove(option)
@@ -169,7 +194,7 @@ class MultipleSelect(Field):
             raise ValueError(
                 "Option(s) %r not found (from %s)"
                 % (', '.join(str_values),
-                   ', '.join([repr(o) for o, c in self.options])))
+                   ', '.join([repr(o) for o, c, t in self.options])))
 
     def value__get(self):
         selected_values = []
@@ -178,7 +203,7 @@ class MultipleSelect(Field):
                                for i in self.selectedIndices]
         elif not self._forced_values:
             selected_values = []
-            for option, checked in self.options:
+            for option, checked, text in self.options:
                 if checked:
                     selected_values.append(option)
         if self._forced_values:
@@ -197,7 +222,7 @@ class Radio(Select):
         if self.selectedIndex is not None:
             return self.options[self.selectedIndex][0]
         else:
-            for option, checked in self.options:
+            for option, checked, text in self.options:
                 if checked:
                     return option
             else:
@@ -413,7 +438,8 @@ class Form(object):
                         assert isinstance(field,
                                           self.FieldClass.classes['radio'])
                     field.options.append((attrs.get('value'),
-                                          'checked' in attrs))
+                                          'checked' in attrs,
+                                          None))
                     continue
                 elif tag_type == 'file':
                     if 'value' in attrs:
@@ -426,7 +452,8 @@ class Form(object):
             if tag == 'select':
                 for option in node('option'):
                     field.options.append((option.attrs.get('value', option.text),
-                                          'selected' in option.attrs))
+                                          'selected' in option.attrs,
+                                          option.text))
 
         self.field_order = field_order
         self.fields = fields
@@ -516,11 +543,34 @@ class Form(object):
             return self[name]
         return fields[index]
 
-    def select(self, name, value, index=None):
-        """Like ``.set()``, except also confirms the target is a ``<select>``.
+    def select(self, name, value=None, text=None, index=None):
+        """Like ``.set()``, except also confirms the target is a ``<select>``
+        and allows selecting options by text.
         """
         field = self.get(name, index=index)
         assert isinstance(field, Select)
+
+        if value is not None and text is not None:
+            raise ValueError("Specify only one of value and text.")
+
+        if text is not None:
+            value = field.get_value_for_text(text)
+
+        field.value = value
+
+    def select_multiple(self, name, value=None, texts=None, index=None):
+        """Like ``.set()``, except also confirms the target is a
+        ``<select multiple>`` and allows selecting options by text.
+        """
+        field = self.get(name, index=index)
+        assert isinstance(field, MultipleSelect)
+
+        if value is not None and texts is not None:
+            raise ValueError("Specify only one of value and texts.")
+
+        if texts is not None:
+            value = field.get_value_for_texts(texts)
+
         field.value = value
 
     def submit(self, name=None, index=None, value=None, **args):